requirements/dev.txt
# Include base requirements
-r base.txt

# Debug toolbar for Django
django-debug-toolbar==4.4.6

# Enhanced Python REPL
ipython==8.30.0

# Testing framework
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-xdist==3.6.1

# Factory for test data generation
factory-boy==3.3.1

# Faker for realistic test data
Faker==33.1.0

# Code formatting and linting
black==24.10.0
flake8==7.1.1
isort==5.13.2

# Type checking
mypy==1.14.0
django-stubs==5.1.1
djangorestframework-stubs==3.15.1

# Documentation generator
sphinx==8.1.3
sphinx-rtd-theme==3.0.2

# API testing
httpie==3.2.4
